+++ CHANGELOG.md
@@ -7,6 +7,7 @@
    - Some spells reward SP's for successful spell casts
    - SP's can be spent to purchase new spells
  - Add alternate_up_parameters, alternate_down_parameters and alternate_sneak_parameters.
+ - Store player EXP and level in backup data, in case of server crash with active wand
 
 ## 5.3